Move common migration functions in to a MicrateCommand class.
